Sprint 16 - moving classes
