--- app/src/main/java/com/example/playlistmaker/data/dto/Response.kt
@@ -1,5 +1,0 @@
-package com.example.playlistmaker.data.dto
-
-open class Response() {
-    var resultCode = 0
-}
