+++ app/src/main/java/com/example/playlistmaker/domain/impl/PlayerInteractorImpl.kt
@@ -2,7 +2,7 @@ package com.example.playlistmaker.domain.impl
 
 import com.example.playlistmaker.domain.api.PlayerInteractor
 import com.example.playlistmaker.domain.enums.PlayerStatus
-import com.example.playlistmaker.domain.models.ITunesTrack
+import com.example.playlistmaker.search.domain.models.ITunesTrack
 import com.example.playlistmaker.domain.repository.PlayerRepository
 
 class PlayerInteractorImpl (private val repository: PlayerRepository): PlayerInteractor {
