+++ app/src/main/java/com/example/playlistmaker/domain/repository/TrackInfoRepository.kt
@@ -1,6 +1,6 @@
 package com.example.vehicle_shop_clean.domain.repository
 
-import com.example.playlistmaker.domain.models.ITunesTrack
+import com.example.playlistmaker.search.domain.models.ITunesTrack
 
 interface TrackInfoRepository {
     fun getTrack(): ITunesTrack
