+++ app/src/main/java/com/example/playlistmaker/search/data/dto/ITunesTrackDto.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.data.dto
+package com.example.playlistmaker.search.data.dto
 
 data class ITunesTrackDto(
     val trackId: Long,
