+++ app/src/main/java/com/example/playlistmaker/search/data/dto/ITunesTrackSearchRequest.kt
@@ -1,3 +1,3 @@
-package com.example.playlistmaker.data.dto
+package com.example.playlistmaker.search.data.dto
 
 data class ITunesTrackSearchRequest(val expression: String)
