+++ app/src/main/java/com/example/playlistmaker/search/domain/api/TracksInteractor.kt
@@ -1,6 +1,6 @@
 package com.example.playlistmaker.search.domain.api
 
-import com.example.playlistmaker.domain.models.ITunesTrack
+import com.example.playlistmaker.search.domain.models.ITunesTrack
 
 interface TracksInteractor {
     fun searchTracks(expression: String, consumer: TracksConsumer)
