+++ app/src/main/java/com/example/playlistmaker/search/presentation/SearchActivity.kt
@@ -15,7 +15,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.playlistmaker.R
 import com.example.playlistmaker.databinding.ActivitySearchBinding
-import com.example.playlistmaker.domain.models.ITunesTrack
+import com.example.playlistmaker.search.domain.models.ITunesTrack
 import com.example.playlistmaker.domain.enums.TrackListStatus
 import com.example.playlistmaker.search.data.SearchHistory
